+++ kernelcache/mcv/pkg/accelerator/devices/nvml.go
@@ -25,7 +25,7 @@ import (
 	"github.com/NVIDIA/go-nvml/pkg/nvml"
 	logging "github.com/sirupsen/logrus"
 
-	"github.com/kserve/kserve/mcv/pkg/config"
+	"github.com/kserve/kserve/kernelcache/mcv/pkg/config"
 )
 
 const (
