+++ kernelcache/mcv/pkg/accelerator/devices/rocm.go
@@ -25,8 +25,8 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/kserve/kserve/mcv/pkg/config"
-	"github.com/kserve/kserve/mcv/pkg/utils"
+	"github.com/kserve/kserve/kernelcache/mcv/pkg/config"
+	"github.com/kserve/kserve/kernelcache/mcv/pkg/utils"
 
 	logging "github.com/sirupsen/logrus"
 )
